KUDU-2830: [java] Improve RPC traces on RPC timeout
Currently when RPCs timeout a full trace of all of the
sent RPCs is printed. Given many retries can occur
within the timeout period, this trace is often very
large and difficult to read and and understand.

This patch introduces a new trace summary format
for the output that prints the most relevant information
in a compact readable way. Along with readability, this
has the benefit of reducing log churn.

The trace sumary format will be used when the log level
is INFO or higher and the original full trace will be used
when the log level is DEBUG or lower.

The new format style is:
Trace Summary(trace-duration ms): Sent(n), Received(n), Delayed(n), MasterRefresh(n), AuthWait(n), Truncated: ?
 Sent: (server-uuid, [ rpc-method, count ], ...), ...
 Received: (server-uuid, [ rpc-status, count ], ...), ...
 Delayed: (server-uuid, [ rpc-method, count ], ...), …

Change-Id: I725e3e8f89f5a433a0ffcb4b0facf82d9ffc4840
Reviewed-on: http://gerrit.cloudera.org:8080/13509
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>