authz: refactor authorization for ListTables
Authorization for ListTables isn't very performant because it requires
authorizing each table while holding its table lock. This invariant is
held to guarantee that the authorization applies to the correct table in
Kudu (e.g. even amidst a concurrent rename).

If we authorize multiple tables at once, we can improve performance, but
keeping the guarantee is less straightforward. This patch takes a stab
at this by refactoring ListTables to prepare for the following
authorization sequence:

1. With table locks held, put together a map from
   { table_name => TableInfo }.
2. Authorize the tables (potentially in bulk in the future), keeping
   track of the authorized table names.
3. Iterate through the authorized table names and, using the maps from
   Step 1, with table locks held, check that the table name hasn't
   changed. If it has, this implies there was a concurrent rename (or
   equivalent) and the authorization call was not actually made for this
   table, and it shouldn't be returned.
4. With confirmation that the tables that were authorized have the IDs
   we expect, return the authorized tables.

If the catalog manager isn't set up to authorize tables (e.g.
authorization isn't enabled), steps 2 and 3 are skipped. This patch
refactors ListTables to match this sequence of events without bulk
authorization.

While this is mainly a refactor, the user-facing change is that if there
is a concurrent rename during a ListTables operation, there is a chance
that neither the old nor new table will show up in the listed tables.
This isn't ideal UX, but is conservative w.r.t security. A test is added
exercising this scenario.

Change-Id: I568e1be7b909768a99f0f4f13775e34a01ccd160
Reviewed-on: http://gerrit.cloudera.org:8080/13566
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Hao Hao <hao.hao@cloudera.com>