thirdparty: fix memkind build race
On rare occasions, the libmemkind.a produced by the memkind build was
missing its jemalloc symbols. This then led to Kudu linkage failures if Kudu
was using static linkage (i.e. in RELEASE builds).

This turned out to be due to a build race of our own making: we can't ask
make to run the install-exec and static_lib targets simultaneously. Both
targets depend on libmemkind.a, but static_lib performs some important
post-processing on libmemkind.a that must complete before install-exec
copies libmemkind.a into the installed directory. And by offering both
targets on the command line, make can run them in parallel once their common
dependency (libmemkind.a) is built.

Change-Id: I1242e733e7d282d1e7471656ea81349be427aa1f
Reviewed-on: http://gerrit.cloudera.org:8080/13513
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins