KUDU-2670 Disable rowset compact for test of SplitKeyRange
We found that in the SplitKeyRange related test, in most cases,
the number of KeyRange is equal to the number of tablets, but we
expect the number of KeyRange to be greater than the number of tablets.

In the test, the DRS generated by flush MRS will be small, rowset
compact will merge them to a big one. This will reduce the number
of DRS in the tablet. If end up with only one DRS in the tablet,
then the tablet will only have one key range.

So, I added --enable_rowset_compaction=false to disable rowset compact,
ensure that the number of KeyRange is greater than the number of tablets.

Change-Id: I89805a1fe2fd4d6c459e4abc0b1b9ec27c3e9588
Reviewed-on: http://gerrit.cloudera.org:8080/13517
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>