[java] Support table rename between scan token creation and rehydration
Previously, if a scan token was created against table 'foo', and table
'foo' was renamed to table 'bar', rehydrating the scan token into a
scanner would either
- fail if there is no longer a table 'foo', or
- attempt to scan another table if that table had been renamed to 'foo'.
This patch alters how the Java client manages scan tokens. It prefers to
use the table id to identify a table, rather than a table name. This
eliminates the above two problems. The table id has been added to the
scan token PB to enable this.

A follow-up will add support for the same in the C++ client.

Change-Id: Iba5b0ac89b9cf98f4cf37c35d740eef744a85005
Reviewed-on: http://gerrit.cloudera.org:8080/13518
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>