KUDU-2809 (3/6): C++ private API for diff scans
This patch adds just enough of a private diff scan API so that fuzz-itest
can use it. I modeled it after Grant's Java work in commit 5e78e4a5c.

No tests here, but this gets plenty of test coverage in an upcoming
fuzz-itest patch, and I think that's fine given fuzz-itest is the only
thing motivating the C++ API in the first place.

Change-Id: I03bd5e51c553b239e36fbdfc4a94131c9599851e
Reviewed-on: http://gerrit.cloudera.org:8080/13535
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>