ksck_remote-test: make checksum tests more robust
They aren't joining the writing thread in the event of a test failure.

I tried to clean up the synchronization with the writer thread but found it
to be too difficult: we really do need to wait until some writes have
happened (i.e. if safe time is not initialized, we fail abruptly), but we
can also make the tests more robust by increasing scanner_max_wait_ms.

Change-Id: I1822e5b31656acf4dfbe1991cd59fa58bb574c81
Reviewed-on: http://gerrit.cloudera.org:8080/13529
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>
Tested-by: Will Berkeley <wdberkeley@gmail.com>