[hms] Adjust storage handler package
This patch changes the new KuduStorageHandler
package from “org.apache.kudu.hive” to
“org.apache.hadoop.hive.kudu”.

This is being done to ensure the stand-in storage handler
can be a real storage handler when a Hive integration
is added in the future. The “org.apache.hadoop.hive”
package is the standard package all Hive storage
handlers lives under.

Note: For the sake of including this into the Impala
project without breaking things, we temporarily
maintain support for the old storage handler. It will
be remvoed in a follow up commit before a release.

Change-Id: I66b314c1c8b56785005d6e0d8b679e19219494fe
Reviewed-on: http://gerrit.cloudera.org:8080/13540
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Tested-by: Kudu Jenkins