[backup] Fix backup CLI dependencies
This patch fixes the backup dependencies in two ways.

First, it fixes the Gradle handling of `isTool` to ensure
sl4j is included.

Second it breaks out the common
“library” type functionality into a kudu-backup-common
module so it can be shared between the kudu-back and
kudu-backup-tools module. This allows
kudu-backup-tools to shade slf4j and log4j without
affecting the kudu-backup dependencies.

I tested this by running a build and then running the
kudu-backup-tools localy via:
   `java -cp build/libs/kudu-backup-tools-1.10.0-SNAPSHOT.jar org.apache.kudu.backup.KuduBackupCLI`

The patch also sneaks in some interface annotations.

Change-Id: I6bee1f467d5652ced8412eef0315057e5b0d684e
Reviewed-on: http://gerrit.cloudera.org:8080/13547
Tested-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Mike Percy <mpercy@apache.org>