KUDU-2841. Expose HMS DB UUID to clients in HiveMetastoreConfig
This fills in the hms_uuid field in HiveMetastoreConfig by fetching the
UUID from the HMS when the HMSCatalog starts up. The new API call is not
available in Hive 2.3 (in our thirdparty build) but has been backported
in some vendor distros and is available upstream in Hive 3.x. In order
to support it where available, this patch manually adds the call to the
thrift definition file, and catches the UNKNOWN_METHOD error code that
comes back when running against Hive 2.3.

I manually tested this against a Hive 3 master and verified that the
VLOG message at startup properly indicated the UUID of the HMS.

Change-Id: I393df4dd281c3fb63a836db346d7b33fb6edc6a2
Reviewed-on: http://gerrit.cloudera.org:8080/13556
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>