[java] Add unit tests demonstrating good behavior of scan tokens when range partitions are dropped
It turns out that scan tokens in the Java client behave nicely with
respect to dropping range partitions- one can drop a range partition
anytime after creating the tokens and the tokens can still be used.
There is a case where some tablets in the dropped partition will still
return rows, while others don't, but that can be looked at as a
manifestation of the fact that tablet alters aren't linearizable in Kudu
(yet?).

Change-Id: Ib25ea2b62321934313c9c8852f0bc00b28bf7f2b
Reviewed-on: http://gerrit.cloudera.org:8080/13565
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Kudu Jenkins