thirdparty: fix build_curl with unusual krb5-config location
The curl build system runs krb5-config to determine which GSSAPI headers
and libraries are needed at build time. It expects to find krb5-config in
/usr/bin and if it's not found, it'll default to linking with -lgssapi.

On SLES12, krb5-config is in /usr/lib/mit/bin, so it isn't found and the
curl build generally fails because -lgssapi isn't a Kudu requirement. By
setting KRB5CONFIG we can redirect curl's build towards the correct location
of krb5-config and get the correct libraries as a result (-lkrb5,
-lk5crypto, and -lcom_err, if you're curious).

Change-Id: Ief5816d561545945089108e34bd59c8e5dae9dc4
Reviewed-on: http://gerrit.cloudera.org:8080/13577
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins