[build] Fix jar verification via verify_jars.pl
This patch fixes jar verification with a few changes.
It fixes some dependencies in kudu-backup and
kudu-test-utils to be sure we aren’t pulling in
unwanted jars. It also updates verify_jars.pl to allow
various relocated classes and the Spark
DataSourceRegister file. It also ignores the buildSrc,
kudu-jepsen and kudu-backup-tools jars.

Change-Id: I57272ff8c1db043c7be00f7f65c11d010b33c8b5
Reviewed-on: http://gerrit.cloudera.org:8080/13584
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Grant Henke <granthenke@apache.org>