KUDU-2823 Place tablet replicas based on dimension
When we add a new range to the fact table, we expect replicas
of the newly created tablet to be evenly distributed across all
available tablet servers.

This is especially important when we use time as the range key.
The more recent the data, the hotter it gets. We expect hot tablets
on the cluster to be evenly distributed.

Unfortunately, after we add some new tablet servers to the cluster,
creating a new tablet replica will prioritize the new tablet server for
placement according to the current placement policy. This is because
we prefer to select tablet server which a smaller number of tablet
replicas. This will cause hot tablets to be concentrated on these new
tablet servers.

So, I added a new placement policy. When creating a new tablet replica,
we prefer to select tablet server which a smaller number of tablet
replicas in the dimension. You can set dimensions when creating tables
and adding partitions, this will ensure that the new tablets are evenly
distributed in the cluster based on dimension.

Change-Id: I48a225e221eb42ef2f5489687e80a151d8dc1a42
Reviewed-on: http://gerrit.cloudera.org:8080/13632
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>