KUDU-2850: fix thrift comparison
In the operator< implementations, we previously wouldn't return early
when determining that a field in the LHS was greater than that in the
RHS, when we should have returned false.

I added a test for each operator I updated that would fail without this
patch.

Change-Id: I071ccb08a385d6a4bfe407ab05621afe437bc29c
Reviewed-on: http://gerrit.cloudera.org:8080/13608
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>
(cherry picked from commit efb648e913452d02496d3cf63f843b8795b7dc12)
Reviewed-on: http://gerrit.cloudera.org:8080/13636