[backup] Fix backup tools packaging
This patch adjusts how we package the kudu-backup-tools
module. Instead of building a fat jar that contains all of
the hadoop classes, the hadoop classes are now marked
as provided and therfore not included.

This change was made with two concerns in mind.
Primarily because the defacto way to load the
environments hadoop configuration is from the
classpath via `hadoop classpath`. A secondary benefit
is that loading the classpath this way avoids any
potential hadoop compatibility issues.

An example execution of the jar would now be:
`java -cp $(hadoop classpath):kudu-backup-tools-1.10.0.jar org.apache.kudu.backup.KuduBackupCLI`

Additionally this patch adds integration tests for the
`hdfs` uri in the tests using a hadoop-minicluster.

Change-Id: I334a38bb5d7252b5e6d2722c4f1fc1d3b5622db7
Reviewed-on: http://gerrit.cloudera.org:8080/13691
Tested-by: Kudu Jenkins
Reviewed-by: Mike Percy <mpercy@apache.org>