[backup] Deflake TestKuduBackup.testRandomBackupAndRestore
TestKuduBackup.testRandomBackupAndRestore creates a random table with
random partitions. In a run with TSAN binaries, I saw the following
partition schema:

partition_schema {
  hash_bucket_schemas {
    columns {
      name: "int160"
    }
    num_buckets: 8
    seed: 395460932
  }
  hash_bucket_schemas {
    columns {
      name: "int321"
    }
    num_buckets: 6
   seed: 2072428334
  }
  hash_bucket_schemas {
    columns {
      name: "int82"
    }
    num_buckets: 9
    seed: 138096156
  }
}

So there were 8 * 6 * 9 = 432 tablets being created on a 3-tserver
minicluster running TSAN. As a result, the test timed out waiting for
the table to finish creating.

To prevent creating tables with too many tablets, this patch reduces the
maximum number of hash buckets to 3 and it reduces the maximum nesting of
hash partitioning to 2. Combined with the existing maximum of 8 range
partitions, this means no table will consist of more than 9 * 8 = 72
tablets. Hopefully this is few enough that create table will always
finish in a timely manner.

Change-Id: Id937f4513bb0ebaf6dec345a141ed9aef2c44388
Reviewed-on: http://gerrit.cloudera.org:8080/13693
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Kudu Jenkins