thirdparty: fix build-if-necessary in tarballs
build-if-necessary, when run from a tarball, uses a build stamp file for
each build configuration to know whether it needs to re-run. However,
a normal build will build both 'common' and then 'tsan' configurations,
in that order. When we go back to check whether 'common' needs a
rebuild, we'll see the tsan build-stamp file and think it needs to be
rebuilt.

This fixes the check to exclude other build-stamp files.

Change-Id: Ifc600d065362e902f4f768080e1f91c90b9f0594
Reviewed-on: http://gerrit.cloudera.org:8080/13706
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Grant Henke <granthenke@apache.org>
Tested-by: Kudu Jenkins