KUDU-2212 mute TSAN on destruction of a locked mutex in glog
This patch adds a suppression for TSAN warning about destruction
of a locked mutex in libglog.  Since the only registered instances
of the issue happen only during shutdown of the C++ runtime upon exit
of the kudu CLI tool, it seems safe to mute it.

The motivation for this change is to hunt down the issue tracked
by KUDU-2878: while doing so, this benign TSAN warning hinders the
scenario used to RCA the latter.

The suppression of the warnings for a locked mutex relies on the fix for
issue 944 in the thread sanitizers. The fix is back-ported as a patch
for LLVM 6.0 in Kudu's thirdparty.  For details on the TSAN suppressions
fix, see https://github.com/google/sanitizers/issues/944.

The results of two representative 1K runs are below: kudu-admin-test built
in TSAN with --gtest_filter='*ToolTest.TestDeleteTable*' before and after
the fix (yes, I ran many of 1K runs for both versions).

  before: 1 out of 1K failed:
    http://dist-test.cloudera.org/job?job_id=aserbin.1565931838.118516

  after : 0 out of 1K failed:
    http://dist-test.cloudera.org/job?job_id=aserbin.1565931149.115149

Change-Id: If0b62027a1a4db18bfaa6c89b3070c918fc4f88a
Reviewed-on: http://gerrit.cloudera.org:8080/13763
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins