Remove experimental tags from NVM block cache gflags
The block cache has supported non-volatile memory for almost four years
now, though one needed to configure some experimental gflags in order to
use it. The implementation was recently changed to use the newer memkind
library and, in the process, underwent additional testing on NVM hardware.
As such, I think it's time to remove the experimental tags to communicate to
users that the underlying code is supported upstream.

Note: I'm not marking them as 'stable' so that we retain the flexibility to
modify the configuration surface itself, if we find we need that.

Change-Id: Ie6b58118116ecc11b2eaa5cec4a9e72c29ac27f8
Reviewed-on: http://gerrit.cloudera.org:8080/13830
Reviewed-by: Greg Solovyev <gsolovyev@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins