[docs] Update the site build to use the release directory
Updates the make_site.sh script and make_docs.sh script
to generate docs in the versioned release directory. This
matches the new symbolic linking style we have been
using for the docs and avoids the need to manually copy
the docs during release.

An additional flag to skip updating the symbolic links (`—no-link`) was added.

Change-Id: I174de2d7041bba20cdc3dcacd3efbb5e736ecb63
Reviewed-on: http://gerrit.cloudera.org:8080/13835
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>