KUDU-2635: ignore failures to delete orphaned blocks
It was previously possible to leave some orphaned blocks in the
in-memory orphaned blocks list by failing to delete some blocks due to a
disk failure. Upon deleting a tablet with such orphaned blocks, Kudu
could crash, as this breaks the assumption in the TabletMetadata that
once we've made the call to delete orphaned blocks, then the orphaned
blocks list will be empty.

This patch addresses this by removing the blocks from the orphaned
blocks list regardless of whether the blocks were deleted at the block
manager level. At worst, this leaves us with untracked blocks, but it's
better than crashing due to a bogus assumption.

A test is added that passed 1/100 times when run on dist-test in debug
mode with stress. With the patch, it passes 100/100 times.

Change-Id: Ice78f41d6d367d42ad31c2127ceb5fc57a244e34
Reviewed-on: http://gerrit.cloudera.org:8080/13858
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>