KUDU-2897 Making the -mavx2 flag contingent on the architecture rather than the OS alone
Earlier, we would use the -mavx2 flag if the OS was UNIX based. However, the flag is only
usable if the underlying CPU architecture supports it. I have added a check to do that.
The following commit has been tested on x86 and ARM processors - one supporting AVX2 and
the other does not. Both were running LINUX.

Change-Id: Ica8c7553cec1fc851685c4a10750e90b3e892500
Reviewed-on: http://gerrit.cloudera.org:8080/13884
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>