version_util: allow period as delimiter in extra and tighten version syntax
This will permit versions like <major>.<minor>.<maintenance>.<vendor_ver>,
a common pattern used by at least one Kudu vendor. The change shouldn't have
any tangible effect upstream as we've always ignored the value of 'extra'.

The string splitting wasn't cutting it for me so I rewrote the parsing to
use POSIX extended regexes. The major side effect is that we're no longer
loosey goosey with components (i.e. "1. 2  .  3" or "+1.+2.+3"); each must
be a subsequence of nothing but digits. I don't think anyone was relying on
this excessive permissiveness and I find the new code to be more readable.

Change-Id: I5502efa8ba378d96432e530366e74a82c83621f7
Reviewed-on: http://gerrit.cloudera.org:8080/13897
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins