version_util: track extra delimiter too
One of the version_util tests verifies that the current version's canonical
representation matches its raw version. The test fails when the delimiter
between the maintenance and extra components is a period, because the
canonical representation builder (i.e. ToString()) assumes that the
delimiter is always a dash.

Rather than weaken the invariant enforced by the unit test, let's track the
parsed delimiter faithfully. I chose to store it separately from the extra
component itself since it didn't logically seem like part of it.

Change-Id: I0fe983e55ae9f83d075297016ccc5d50c208c549
Reviewed-on: http://gerrit.cloudera.org:8080/13911
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>