[thirdparty] introduce chrony
Added chrony into the thirdparty: a set of follow-up patches use the
chronyd NTP implementation and its chronyc CLI to provide the
functionality of reference NTP servers for test scenarios involving
Kudu built-in NTP client.

One nice feature that chronyd has is the ability to run in server-only
mode, i.e. not driving the system clock: 'man chronyd', the '-x' option.
Also, it's possible to chronyd NTP server using the system clock
as a source instead of a physical device (GPS, oscillator, etc.).
In addition, it's possible to manually set the reference time for
chrony NTP server running in server-only local mode.  The chrony suite
has 'chronyc' CLI tool to send control commands to chronyd NTP daemon.

As of now, the latest version 3.5 is used with not-yet-upstream patch
to drop the requirements of being run under superuser UID when driving
the system clock is not required (e.g., when running in server-only
mode, see '-x' command-line flag).

Change-Id: I71ed12311b10979af8a12094881b6b8b47ef8008
Reviewed-on: http://gerrit.cloudera.org:8080/13915
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>