KUDU-2881 Support create/drop range partition by command line
Sometimes we need to drop the range partition and then recreate it in
order to rewrite in case that the partition was written wrongly. This
patch supports to add/drop range partition by command line. The command
can be used as:
1. kudu table add_range_partition <master_addresses> <table_name>
<lower_bound> <upper_bound> [-lower_bound_type] [-upper_bound_type]
2. kudu table drop_range_partition <master_addresses> <table_name>
<lower_bound> <upper_bound> [-lower_bound_type] [-upper_bound_type]

The parameters of lower_bound and upper_bound can be empty, which means
we will use the default value. If both these two parameters are empty,
it means the range partition is unbounded.

The parameters of lower_bound_type and upper_bound_type is optional. If
these two parameters are not specified, the default value are
INCLUSIVE_BOUND and EXCLUSIVE_BOUND respectively. These two parameters
are both case-insensitive.

Change-Id: I647d9c3cf01807bd8ae9f8cf4091568ea6f1161c
Reviewed-on: http://gerrit.cloudera.org:8080/13833
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins