KUDU-2907: add a dir to dir groups when full
This patch extends the directory-picking behavior when selecting
directories for new blocks if all directories in the given tablet's
group are full. It now allows the expansion of a tablet's data dir group
if healthy directories are available.

This doesn't change the policy for picking directories for groups; it
only extends the usage of the policy to potentially add directories when
getting a directory for a new block.

Change-Id: Ic8e05bdfe1fc2ac0e7152d493f289d3ac6e850d7
Reviewed-on: http://gerrit.cloudera.org:8080/14122
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>