[master] update tags for --master_client_location_assignment_enabled
The --master_client_location_assignment_enabled flag was originally
purposed for test scenarios.  However, there is nothing unsafe
in assigning locations only to tablet servers, but not clients.

This patch removes 'unsafe' tag from the flag and adds 'advanced' tag
instead.  In addition, this patch cleans up the related code a bit,
adding TODO for easier tracking of further improvements
in the scope of KUDU-2771.

Change-Id: Iaeb7391174a37f1a774f83b964bee854922cadcd
Reviewed-on: http://gerrit.cloudera.org:8080/14190
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins