KUDU-2069 p1: add persistent tserver maintenance mode
Adds tablet server states to the master, with which to represent tablet
server maintenance mode (and in the future, decommissioning).

The on-disk state is represented as an entry in the system catalog
table, keyed by the tserver UUID. The in-memory state is maintained by
the TSManager as a mapping from tserver UUID to the new TServerStatePB
enum.

Note that this patch only introduces the states internally and doesn't
expose a way for users to set them.

Testing:
- test to check that maintenance mode is maintained through a restart of
  the master
- test to ensure that repeated setting of state does not affect the
  in-memory tserver states
- test that concurrently updates tserver state and ensures that the
  results in-memory match with what's on-disk

Change-Id: Ib669b43b3cee171c4c7dbd54041e29c30cb9f767
Reviewed-on: http://gerrit.cloudera.org:8080/14217
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins