KUDU-2069 p3: add RPC endpoint for maintenance mode
This patch adds an RPC endpoint to the master to allow entering and
exiting maintenance mode. The endpoint requires superuser privileges, as
this is meant to be used for administrative purposes.

The endpoint is currently hidden behind a flag, as maintenance mode
behavior is still in progress. I'm introducing the endpoint now to
facilitate testing and will remove the flag once finished.

The requests are represented as explicit edges in a (currently simple)
state machine transitioning a given tserver between different tserver
states.

Basic checking is done to validate requests, and a test is added for
this checking.

Change-Id: I9d565bd745507f2511b91a96d2d446240c5406b5
Reviewed-on: http://gerrit.cloudera.org:8080/14221
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins